% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_colors.R
\docType{data}
\name{cat_colors}
\alias{cat_colors}
\title{20 different color - categorical group. It is useful with work with scatterplot, boxplot or line.
previous name: categorical_colors}
\format{An object of class \code{character} of length 20.}
\usage{
cat_colors
}
\description{
20 different color - categorical group. It is useful with work with scatterplot, boxplot or line.
previous name: categorical_colors
}
\keyword{datasets}
