% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_enr.R
\name{datatable_enr}
\alias{datatable_enr}
\title{Create datatable in ENR style}
\usage{
datatable_enr(data, options = list(), sorting = FALSE, desc_column = NULL,
  tooltip = "", id = NULL, ...)
}
\arguments{
\item{data}{a data object}

\item{options}{a list of initialization options. All things available in DT package.}

\item{sorting}{binary parameter regarding soritng columns.}

\item{desc_column}{describing variables in the table. Numerical argument, e.g. 1:3 treats first three columns as a descr. variables.}

\item{tooltip}{tooltip which descibes content's table.}

\item{id}{table id. It's necessary in shiny app when you want to add a tooltip. Id must match with id of render object. More in examples.}

\item{...}{all DT::datatable parameters (exept options which you can define in options argument).}
}
\description{
Create datatable in ENR style
}
\examples{
\dontrun{
datatable_enr(mtcars, 
        desc_column = 1:3,
        tooltip = "This is a description of table."
)

output$mtcars_table <- renderDataTable({
 
 datatable_enr(mtcars, tooltip = "Lorem", id = "mtcars_table")
 
})
}
}
