% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_colors.R
\docType{data}
\name{div_colors}
\alias{div_colors}
\title{11 different color - diverging group (from Red to Blue). It is useful with work with barplot or mapplot.
previous name: diverging_colors}
\format{An object of class \code{character} of length 11.}
\usage{
div_colors
}
\description{
11 different color - diverging group (from Red to Blue). It is useful with work with barplot or mapplot.
previous name: diverging_colors
}
\keyword{datasets}
