% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_scale_manual.R
\name{draw_scale_manual}
\alias{draw_scale_manual}
\title{Change color from categorical colors to diverging/sequential colors}
\usage{
draw_scale_manual(method = "", color = cat_colors,
  legend_title = "Legend")
}
\arguments{
\item{method}{Which method was used to differ variables in a plot? Two possible value: color or fill.}

\item{color}{Which color group we want to implement? Four possible group: cat_colors,div_colors, seq_colors or other color vector. Default value: cat_colors}

\item{legend_title}{Title of legend. Default value: "Legend"}
}
\description{
Change color from categorical colors to diverging/sequential colors
}
\examples{
\dontrun{
ggplot(mtcars, aes(factor(cyl), wt))+
     geom_bar(stat = "identity", fill = cat_colors[1])+
     theme_peru()+
     draw_scale_numeric("y",0,60)

ggplot(diamonds[1:100,], aes(depth, carat, color = factor(cut)))+
     geom_point(size = 7)+
     theme_peru("color")+
     draw_scale_numeric("y",0,1,0.1,labels_percent = T)
     }
}
