% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_scale_numeric.R
\name{draw_scale_numeric}
\alias{draw_scale_numeric}
\title{Manipulate value of text in given axis.}
\usage{
draw_scale_numeric(axis, min, max, step = (max - min)/10,
  labels = scales::comma)
}
\arguments{
\item{axis}{Name of axis - "x" or "y"}

\item{min}{Minimal value of given axis.}

\item{max}{Maximal value of given axis.}

\item{step}{How frequently we show text in given axis.
Default value is based on max and min arguments: (max - min)/10.}

\item{labels}{Show value in different way. Default value: comma;
other options: percent.}
}
\description{
Manipulate value of text in given axis.
}
\examples{
\dontrun{
ggplot(mtcars, aes(factor(cyl), wt))+
     geom_bar(stat = "identity", fill = cat_colors[1])+
     theme_peru()+
     draw_scale_numeric("y",0,60)

ggplot(diamonds[1:100,], aes(depth, carat, color = factor(cut)))+
     geom_point(size = 7)+
     theme_peru("color")+
     draw_scale_numeric("y",0,1,0.1,labels = percent)
     }
}
