% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_data.R
\name{get_html_data}
\alias{get_html_data}
\title{Create HTML format data. Useful with datable_enr function.}
\usage{
get_html_data(data, column = NULL, direction = "col",
  formatter = "simple_formatter")
}
\arguments{
\item{data}{dataset you want to implement in DT::datatable.}

\item{column}{columns which will be format by 'formatter' argument. Those columns will be centered.}

\item{direction}{the direction in which cells will be coloured. Possible opitons: 'area', 'col' and 'row'. Default value: 'col'.}

\item{formatter}{method of color choosen columns' cells. Default value: 'custom_formatter'.}
}
\description{
Create HTML format data. Useful with datable_enr function.
}
\examples{
\dontrun{
example_dt <- get_html_data(mtcars, column = 4:6, direction = 'area')
datatable_enr(example_dt, tooltip = "This is a description.")
}
}
