% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_enr.R
\name{legend_enr}
\alias{legend_enr}
\title{E&R legend - put legend inside a plot.}
\usage{
legend_enr(geom = "bar", ...)
}
\arguments{
\item{geom}{Name of used geom. Default value: bar.}

\item{...}{arguments listed below:}

\item{direction}{direction of the legend. Either h (horizontal) or v (vertical).}

\item{direction_box}{direction of the box of legend. Specifically it is useful when the plot has more than one legend. Default value: v.}

\item{position}{position of the legend - two-element numeric vector.}

\item{bc}{- background color. Default value: #FFFFFF.}

\item{bc_alpha}{- opacity of background color. Default value: 0.}

\item{bc_border}{- color of border of legend border. Default value: transparent.}

\item{lo_size}{size of legend objects. Default value for geom_point: 8, for geom_line: 1.2}

\item{lo_alpha}{opacity of legend objects. Default value: 1}

\item{lo_nrow}{number of rows in which categories should show up.}
}
\description{
E&R legend - put legend inside a plot.
}
\examples{
\dontrun{
ggplot(mtcars, aes(qsec, mpg, color = factor(cyl))) +
    geom_point() +
    theme_enr() +
    legend_enr("point")
     }
}
