% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_enr_fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{Download OpenSans fonts}
\usage{
load_fonts(message = TRUE)
}
\arguments{
\item{message}{message that all necessary fonts are installed. Default value: TRUE}
}
\description{
Download OpenSans fonts
}
