% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_colors.R
\docType{data}
\name{p_colors}
\alias{p_colors}
\title{8 different colors - Pearson colors (second palette)}
\format{An object of class \code{character} of length 8.}
\usage{
p_colors
}
\description{
8 different colors - Pearson colors (second palette)
}
\keyword{datasets}
