% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_colors.R
\docType{data}
\name{pair_colors}
\alias{pair_colors}
\title{6 pair colors -  It is useful with work with scatterplot, boxplot or line.}
\format{An object of class \code{character} of length 12.}
\usage{
pair_colors
}
\description{
6 pair colors -  It is useful with work with scatterplot, boxplot or line.
}
\keyword{datasets}
