% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quar_colors.R
\docType{data}
\name{quar_colors}
\alias{quar_colors}
\title{4 quartet colors -  It is useful with work with scatterplot, boxplot or line.}
\format{An object of class \code{character} of length 16.}
\usage{
quar_colors
}
\description{
4 quartet colors -  It is useful with work with scatterplot, boxplot or line.
}
\keyword{datasets}
