% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Function for saving a plot.}
\usage{
save_plot(plot_name, file_name = "plot", type_file = "png", width = 13,
  height = 8, dpi = 150)
}
\arguments{
\item{plot_name}{Name of plot we want to export.}

\item{file_name}{Name of file we want to see in directory. Default filename: "plot"}

\item{type_file}{Type of file we want to export. It is possible to export plot in  certain format:eps/ps, pdf, jpeg, tiff, png, bmp, svg Default type of file: png}

\item{width}{Width of exported plot. Default value: 13 inches}

\item{height}{Height of exported plot. Default value: 8 inches}

\item{dpi}{Default value: dpi = 150}
}
\description{
Function for saving a plot.
}
