% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot2.R
\name{save_plot2}
\alias{save_plot2}
\title{Function for saving a plot in pdf and png format at once.}
\usage{
save_plot2(name, file = "plot", width = 13, height = 8, dpi = 150)
}
\arguments{
\item{name}{Plot name we want to export.}

\item{file}{File name or path. Default filename: "plot"}

\item{width}{Width of exported plot. Default value: 13 inches}

\item{height}{Height of exported plot. Default value: 8 inches}

\item{dpi}{Default value: dpi = 150}
}
\description{
Function for saving a plot in pdf and png format at once.
}
