% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_grid.R
\name{save_plot_grid}
\alias{save_plot_grid}
\title{Function for saving a plot (when it is not ggobject).}
\usage{
save_plot_grid(plot_name, file_name = "plot", width = 13, height = 8,
  dpi = 150)
}
\arguments{
\item{plot_name}{Name of plot we want to export.}

\item{file_name}{Name of file we want to see in directory. Default filename: "plot"}

\item{width}{Width of exported plot. Default value: 13 inches}

\item{height}{Height of exported plot. Default value: 8 inches}

\item{dpi}{Default value: dpi = 150}
}
\description{
Function for saving a plot (when it is not ggobject).
}
