% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_colors.R
\docType{data}
\name{seq_colors}
\alias{seq_colors}
\title{9 different colors - sequential group (Blue gradient). It is useful with work with barplot or mapplot.
previous name: sequential_colors}
\format{An object of class \code{character} of length 9.}
\usage{
seq_colors
}
\description{
9 different colors - sequential group (Blue gradient). It is useful with work with barplot or mapplot.
previous name: sequential_colors
}
\keyword{datasets}
