% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_data.R
\name{simple_formatter}
\alias{simple_formatter}
\title{formatter no 2}
\usage{
simple_formatter(min.col = "#D4EAE4", max.col = "#007FA3")
}
\arguments{
\item{min.col}{Minimum of gradient color. By default. "#D4EAE4"}

\item{max.col}{Maximum of gradient color. By default. "#007FA3"}
}
\description{
formatter no 2
}
