% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio_colors.R
\docType{data}
\name{trio_colors}
\alias{trio_colors}
\title{6 trio colors -  It is useful with work with scatterplot, boxplot or line.}
\format{An object of class \code{character} of length 18.}
\usage{
trio_colors
}
\description{
6 trio colors -  It is useful with work with scatterplot, boxplot or line.
}
\keyword{datasets}
