% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vispear.R
\docType{package}
\name{vispear-package}
\alias{vispear}
\alias{vispear-package}
\title{vispear: A package facilitating data visualisation using Pearson branding}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pearsonplc/vispear}
  \item Report bugs at \url{https://github.com/pearsonplc/vispear/issues}
}

}
\keyword{internal}
