% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_colors.R
\docType{data}
\name{p_colors}
\alias{p_colors}
\title{4 different colors - Pearson colors}
\format{An object of class \code{character} of length 4.}
\usage{
p_colors
}
\description{
4 different colors - Pearson colors
}
\keyword{datasets}
