% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_data.R
\name{simple_formatter}
\alias{simple_formatter}
\title{formatter no 2}
\usage{
simple_formatter(min.col = p_colors[4], max.col = p_colors[1])
}
\arguments{
\item{min.col}{Minimum of gradient color. By default. vispear::p_colors[4]}

\item{max.col}{Maximum of gradient color. By default. vispear::p_colors[1]}
}
\description{
formatter no 2
}
