% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vispear.R
\docType{package}
\name{vispear-package}
\alias{vispear}
\alias{vispear-package}
\title{vispear: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{}
}

}
\keyword{internal}
